Fix issues in togge nav
